ensure we don't attempt to inspect the heap for a non-existent key

+++ lib/Autocache/Store/MemoryLRU.pm
@@ -94,6 +94,7 @@ sub delete
     my ($self,$key) = @_;
     get_logger()->debug( "delete: $key" );
     my $elem = delete $self->_cache->{$key};
+    return unless $elem;
     $self->_heap->delete( $elem );
     $self->size( $self->size - $elem->val->size );
     return $elem->val->val;
